+++ src/fhir/R4R5-common/sdc-import.js
@@ -13,8 +13,7 @@ function addSDCImportFns(ns) {
   var self = ns;
 
   // FHIR extension urls
-  self.fhirExtUrlOptionScore = "http://hl7.org/fhir/StructureDefinition/ordinalValue";
-  self.fhirExtUrlValueSetScore = self.fhirExtUrlOptionScore;
+  self.fhirExtUrlValueSetScore = "http://hl7.org/fhir/StructureDefinition/ordinalValue";
 
 
   /**
@@ -144,11 +143,12 @@ function addSDCImportFns(ns) {
           if(label) {
             answer.label = label.valueString;
           }
-          var score = LForms.Util.findObjectInArray(option.extension, 'url', self.fhirExtUrlOptionScore);
+          // Any of the URLs in self.fhirExtUrlOptionScoreLookup should work on import regardless of the version of FHIR.
+          var score = option.extension?.find(ext => self.fhirExtUrlOptionScoreUrlSet.has(ext.url));
           // Look for argonaut extension.
           score = !score ? LForms.Util.findObjectInArray(option.extension, 'url', self.argonautExtUrlExtensionScore) : score;
           if(score) {
-            answer.score = score.valueDecimal.toString();
+            answer.score = parseFloat(score.valueDecimal);
           }
 
         }